rename dapp-box to conflux-hub
