+++ dapps/conflux-hub/index.html
@@ -6,14 +6,14 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <meta name="description" content="Easily store, send and receive funds on Conflux.Fluent supports HD wallets and hardware wallets." />
     <meta name="keywords" content="Conflux Wallet, Hierarchical Deterministic Wallet">
-    <title>Conflux dapp box</title>
+    <title>Conflux Hub</title>
     <title>Fluent - A simple and secure Conflux wallet built for Web 3</title>
     <script>
         var global = window;
     </script>
   </head>
   <body class="light">
-    <div id="dapp-box-root"></div>
+    <div id="conflux-hub-root"></div>
     <script type="module" src="/src/main.tsx"></script>
   </body>
 </html>
