+++ dapps/conflux-hub/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "dapp-box",
+  "name": "conflux-hub",
   "private": true,
   "version": "0.0.0",
   "scripts": {
