+++ dapps/conflux-hub/src/App.css
@@ -1,9 +1,9 @@
-#dapp-box-root {
+#conflux-hub-root {
     @apply flex;
     font-family: 'SF Pro Display';
 }
 
-#dapp-box-root > .scrollbar__wrapper {
+#conflux-hub-root > .scrollbar__wrapper {
     @apply flex-1 px-[32px];
 }
