+++ dapps/conflux-hub/src/App.tsx
@@ -4,12 +4,12 @@ import CustomScrollbar from 'custom-react-scrollbar';
 import Navbar from 'common/modules/Navbar';
 import { LocaleContext } from 'common/hooks/useI18n';
 import { ModeContext } from 'common/hooks/useMode';
-import Sidebar from 'dapp-box/src/modules/Sidebar';
+import Sidebar from 'conflux-hub/src/modules/Sidebar';
 import CrossSpace from 'cross-space/src/modules';
-import ShuttleFlowNavbarEnhance from 'dapp-box/src/modules/NavbarEnhance/ShuttleFlow';
-import useCurrentDapp from 'dapp-box/src/hooks/useCurrentDapp';
-import ShuttleFlowIcon from 'dapp-box/src/assets/shuttle-flow.svg';
-import CrossSpaceIcon from 'dapp-box/src/assets/cross-space.svg';
+import ShuttleFlowNavbarEnhance from 'conflux-hub/src/modules/NavbarEnhance/ShuttleFlow';
+import useCurrentDapp from 'conflux-hub/src/hooks/useCurrentDapp';
+import ShuttleFlowIcon from 'conflux-hub/src/assets/shuttle-flow.svg';
+import CrossSpaceIcon from 'conflux-hub/src/assets/cross-space.svg';
 import { hideAllToast } from 'common/components/tools/Toast';
 import './App.css';
 import 'common/index.css';
