+++ dapps/conflux-hub/src/vite-env.d.ts
