+++ dapps/cross-space/src/App.tsx
@@ -4,7 +4,7 @@ import Navbar from 'common/modules/Navbar';
 import Modules from 'cross-space/src/modules';
 import { LocaleContext } from 'common/hooks/useI18n';
 import { ModeContext } from 'common/hooks/useMode';
-import CrossSpaceIcon from 'dapp-box/src/assets/cross-space.svg';
+import CrossSpaceIcon from 'conflux-hub/src/assets/cross-space.svg';
 import './App.css';
 
 const AppRouter = () => {
